requirements.txt
Flask
apscheduler
requests
peewee
gunicorn==21.2.0
python-dotenv==1.0.0
openai